add firs test create folding cell
