+++ FoldingCell/FoldingCellDemoTests/Info.plist
